refactor: extract misc.rs, make commands/mod.rs pure pub use hub (#127 step 10)
Move all remaining functions (greet, list_system_monospace_fonts,
settings/memo/clipboard/notification/cache/window-geometry commands,
GitHub issue submission, etc.) to commands/misc.rs.

commands/mod.rs is now a ~12-line pure re-export hub with no logic.
389 tests passing.
